fix visual effect from budgets (client/business) and navigation + question before delete

+++ app/src/main/java/com/edifica/fragments/clients/FragmentClientBudgets.kt
@@ -10,9 +10,12 @@ import android.widget.TextView
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.edifica.R
+import com.edifica.activities.clients.ActivityClientChat
+import com.edifica.activities.clients.ActivityClientMain
 import com.edifica.adapters.TransactionsClientAdapter
 import com.edifica.interfaces.TransactionListener
 import com.edifica.models.*
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.DocumentReference
 import com.google.firebase.firestore.FirebaseFirestore
@@ -22,12 +25,11 @@ import com.google.firebase.firestore.FirebaseFirestore
  */
 class FragmentClientBudgets : Fragment(), TransactionListener {
 
-    var transactions: ArrayList<Transactions> = arrayListOf()
+    private var transactions: ArrayList<Transactions> = arrayListOf()
     var db = FirebaseFirestore.getInstance()
     var auth = FirebaseAuth.getInstance()
-    val TAG = "miApp"
-    var recyclerView: RecyclerView? = null
-    lateinit var mAdapter: RecyclerView.Adapter<TransactionsClientAdapter.MainViewHolder>
+    private var recyclerView: RecyclerView? = null
+    private lateinit var mAdapter: RecyclerView.Adapter<TransactionsClientAdapter.MainViewHolder>
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -101,15 +103,29 @@ class FragmentClientBudgets : Fragment(), TransactionListener {
     }
 
     override fun cancelOnItemClick(transaction: Transactions, position: Int) {
-        db.collection("transaction").document(transaction.id).delete()
 
-        transactions.remove(transaction)
+        val dialogBuilder = MaterialAlertDialogBuilder(context, R.style.MyMaterialAlertDialog)
 
-        loadAdapter()
+        dialogBuilder.setMessage(R.string.fragment_client_profile_chat_question_delete_alert_dialog)
+            .setPositiveButton(R.string.fragment_client_profile_chat_affirmative_delete_alert_dialog) { _, _ ->
+                db.collection("transaction").document(transaction.id).delete()
+
+                transactions.remove(transaction)
+
+                loadAdapter()
+            }
+            .setNegativeButton(R.string.fragment_client_profile_chat_negative_delete_alert_dialog) { dialog, _ ->
+                dialog.cancel()
+            }
+
+        val alert = dialogBuilder.create()
+
+        alert.setTitle(R.string.fragment_client_profile_chat_action_delete_alert_dialog)
+        alert.show()
     }
 
     override fun chatOnItemClick(transaction: Transactions, position: Int) {
-        TODO("Not yet implemented")
+        (activity as ActivityClientMain).gotoActivity(ActivityClientChat())
     }
 
 }
